HMP-347: removes redundant labels from search form
