+++ extras/bampfa/app/views/shared/_header_navbar.html.erb
@@ -12,7 +12,7 @@
 
 <%= content_tag :div, class: 'navbar navbar-search bg-light mb-1', role: 'navigation', aria: { label: t('blacklight.search.header') } do %>
   <div class="<%= container_classes %>">
-    <%= render((blacklight_config&.view_config(document_index_view_type)&.search_bar_component ||Blacklight::SearchBarComponent).new(
+    <%= render((blacklight_config&.view_config(document_index_view_type)&.search_bar_component || SearchBarComponent).new(
           url: search_action_url,
           advanced_search_url: search_action_url(action: 'advanced_search'),
           params: search_state.params_for_search.except(:qt),
