+++ extras/cinefiles/config/locales/blacklight.en.yml
@@ -10,7 +10,7 @@ en:
         title: 'Filter your search'
       form:
         search:
-          label: 'search for'
+          label: 'search for keywords'
           placeholder: 'Enter a few keywords here to get started...'
       start_over: 'Reset'
       pagination_info:
