+++ extras/pahma/app/views/shared/_header_navbar.html.erb
@@ -15,7 +15,7 @@
 
 <%= content_tag :div, class: 'navbar navbar-search navbar-light bg-light mb-1', style:'background-image: url(/pahma_17-379a-c.png); width:100%', role: 'navigation', aria: { label: t('blacklight.search.header') } do %>
   <div class="<%= container_classes %>">
-    <%= render((blacklight_config&.view_config(document_index_view_type)&.search_bar_component ||Blacklight::SearchBarComponent).new(
+    <%= render((blacklight_config&.view_config(document_index_view_type)&.search_bar_component || SearchBarComponent).new(
           url: search_action_url,
           advanced_search_url: search_action_url(action: 'advanced_search'),
           params: search_state.params_for_search.except(:qt),
