+++ portal/app/components/search_bar_component.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class SearchBarComponent < Blacklight::SearchBarComponent
+
+end
