+++ portal/app/views/catalog/_search_form.html.erb
@@ -1,13 +1,10 @@
 <%= warn "#{__FILE__} is a deprecated partial." %>
-<%= render((blacklight_config&.view_config(document_index_view_type)&.search_bar_component || Blacklight::SearchBarComponent).new(
+<%= render((blacklight_config&.view_config(document_index_view_type)&.search_bar_component || SearchBarComponent).new(
       url: search_action_url,
       advanced_search_url: search_action_url(action: 'advanced_search'),
       params: search_state.params_for_search.except(:qt),
       search_fields: Deprecation.silence(Blacklight::ConfigurationHelperBehavior) { search_fields },
       autocomplete_path: search_action_path(action: :suggest))) %>
-
-
-
 <div>
   <%= link_to 'More options', blacklight_advanced_search_engine.advanced_search_path(search_state.to_h), class: 'advanced_search btn btn-secondary'%>
 </div>
